ref(ngrx-ducks): move interface for pick to typings/
