+++ ngrx/ducks/src/lib/core/types/index.ts
@@ -1,8 +1,8 @@
 export * from '../../typings/action-creator-for-effects';
+export * from '../../typings/pick-selector';
 export * from './action-creator/action-creator';
 export * from './case-reducer/case-reducer';
 export * from './ducks/duck';
 export * from './ducks/ducks';
-export * from './ducks/pick-selector';
 export * from './wired-actions/wired-action';
 export * from './wired-actions/wired-actions';
