+++ ngrx/ducks/src/lib/typings/pick-selector.ts
